+++ include/salmon/mesh.h
@@ -9,7 +9,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
-#include <shader.h>
+#include <salmon/shader.h>
 
 #include <string>
 #include <vector>