+++ src/imgui_layer.cpp
@@ -1,8 +1,8 @@
-#include <imgui_layer.h>
+#include <salmon/imgui_layer.h>
 #include <imgui/imgui.h>
 #include <imgui/imgui_impl_glfw.h>
 #include <imgui/imgui_impl_opengl3.h>
-#include <engine.h>
+#include <salmon/engine.h>
 
 namespace ImGuiLayer
 {